+++ scripts/create_training_data.py
@@ -53,8 +53,8 @@ def process_one_file(diagnostics_file, output_filename):
         # Write out matrices with a newaxis at front (for # of channels)
         # Also downcast to float32, to save space and for GPU-ability
         deepdish.io.save(
-            output_filename, {'X': midi_gram[np.newaxis].astype(np.float32),
-                              'Y': audio_gram[np.newaxis].astype(np.float32)})
+            output_filename, {'X': midi_gram[np.newaxis],
+                              'Y': audio_gram[np.newaxis]})
     except Exception as e:
         print "Error for {}: {}".format(
             diagnostics_file, traceback.format_exc(e))
